pyproject.toml
[tool.poetry]
name = "clashprofile"
version = "1.4.0"
description = "通过订阅链接生成clash配置文件。"
authors = ["Well404 <well_404@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^6.0"
uvicorn = "^0.19.0"
fastapi = "^0.86.0"
httpx = "^0.23.0"
APScheduler = "^3.9.1.post1"
loguru = "^0.6.0"
pydantic = "^1.10.2"


[tool.poetry.group.dev.dependencies]
black = {version = "^22.10.0", allow-prereleases = true}

# [[tool.poetry.source]]
# name = "aliyun"
# url = "https://mirrors.aliyun.com/pypi/simple/"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
